+++ source/numeric/ConstantValue.cpp
@@ -652,7 +652,7 @@ ConstantRange ConstantRange::subrange(ConstantRange select) const {
     result.right = select.upper() + l;
 
     SLANG_ASSERT(result.right <= upper());
-    if (isLittleEndian())
+    if (isDescending())
         return result;
     else
         return result.reverse();
@@ -669,7 +669,7 @@ ConstantRange ConstantRange::intersect(ConstantRange other) const {
 }
 
 int32_t ConstantRange::translateIndex(int32_t index) const {
-    if (!isLittleEndian())
+    if (!isDescending())
         return upper() - index;
     else
         return index - lower();
@@ -687,7 +687,7 @@ bool ConstantRange::overlaps(ConstantRange other) const {
     return lower() <= other.upper() && upper() >= other.lower();
 }
 
-std::optional<ConstantRange> ConstantRange::getIndexedRange(int32_t l, int32_t r, bool littleEndian,
+std::optional<ConstantRange> ConstantRange::getIndexedRange(int32_t l, int32_t r, bool descending,
                                                             bool indexedUp) {
     ConstantRange result;
     int32_t count = r - 1;
@@ -708,7 +708,7 @@ std::optional<ConstantRange> ConstantRange::getIndexedRange(int32_t l, int32_t r
         result.right = *lower;
     }
 
-    if (!littleEndian)
+    if (!descending)
         return result.reverse();
 
     return result;